+++ ui/src/store/modules/story/apollo-helpers.js
@@ -1,0 +1,15 @@
+import gql from 'graphql-tag'
+
+export const deleteStory = (id, date) => ({
+  mutation: gql`
+    mutation ($input: StoryInput!) {
+      saveStory (input: $input)
+    }
+  `,
+  variables: {
+    input: {
+      id,
+      tombstone: { date }
+    }
+  }
+})